% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootMargin.R
\name{bootMargin}
\alias{bootMargin}
\title{Significant Mean Causes and Effects For Complete and Chain Bipartite Graphs}
\usage{
bootMargin(
  CC = NULL,
  CE = NULL,
  EE = NULL,
  no.zeros = TRUE,
  thr.cause = 0.5,
  thr.effect = 0.5,
  reps = 10000,
  conf.level = 0.95,
  delete = FALSE,
  plot = FALSE
)
}
\arguments{
\item{CC}{TIt allows for entering a three-dimensional incidence array, with each submatrix
along the z-axis being a square incidence matrix. By default, CC = NULL.}

\item{CE}{It allows for entering a three-dimensional incidence array, with each submatrix
along the z-axis being a square incidence matrix (for complete graphs) or a rectangular
 matrix (for chained bipartite graphs).}

\item{EE}{It allows for entering a three-dimensional incidence array, with each submatrix
along the z-axis being a square incidence matrix. By default, EE = NULL.}

\item{no.zeros}{For no-zeros = TRUE, the function omits zeros in the calculations.
By default, no-zeros = TRUE.}

\item{thr.cause}{Defines the degree of truth in which incidence is considered significant
within the range [0,1]. By default thr.cause = 0.5}

\item{thr.effect}{Defines the degree of truth in which incidence is considered significant
within the range [0,1]. By default thr.effect = 0.5}

\item{reps}{Defines the number of bootstrap replicates. By default, reps = 10000.}

\item{conf.level}{Defines the confidence level. By default, conf.level = 0.95.}

\item{delete}{Removes the non-significant results from the $DirectEffects set and
returns the entered three-dimensional incidence arrays by assigning zeros to the
edges whose incidences are significantly lower than thr at the p-value set in the
conf.level parameter. By default, delete = FALSE.}

\item{plot}{Generates a Dependence-Influence plot with the data from $byCause and $byEffect.
The "Dependence" associated with $byEffect is on the X-axis, and the "Influence"
associated with $byCause is on the Y-axis.}
}
\value{
The function returns a list with subsets of data.
The subset $byRow and $byCol contains the following values:
 \item{Var}{Variable name.}
 \item{Mean}{Calculated mean.}
 \item{LCI}{Lower Confidence Interval.}
 \item{UCI}{Upper Confidence Interval.}
 \item{p.value}{The calculated p-value.}
For delete = TRUE with complete graphs, the function returns $Data, the matrix entered
 in the CC parameter, but with the non-significant rows and columns removed.

For delete = TRUE with chained bipartite graphs, the function returns $CC, $CE, $EE
, the three-dimensional matrices entered in the parameters CC, CE and EE, but
eliminating the non-significant rows and columns.

For plot = TRUE, the function returns the subset $plot, which contains the graph
generated from the data $byCause and $byEffect on the X axis there is "dependency"
associated with $byEffect and on the Y axis "influence" associated with $byCause.
}
\description{
The bootMargin() function calculates the mean incidence of each cause and each effect,
 confidence intervals, and p-value for complete and chained bipartite graphs using multiple key informants.
  This function performs a t-test with bilateral contrast via bootstrap BCa.
}
\details{
bootMargin
}
\examples{
# For chain bipartite graphs the parameters CC, CE and EE are used.
# For instance:
bootMargin(CC = CC, CE = CE, EE = EE, thr.cause = 0.5, thr.effect = 0.5, reps = 500)
}
