% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CC}
\alias{CC}
\title{Public policy data on air pollution in the Valdivia commune, Chile.}
\format{
Three-dimensional matrix with 16 incentives and 10 key experts.
\describe{
  \item{I1}{Prohibition}
  \item{I2}{Standard/regulation}
  \item{I3}{Subsidy for heating devices for public services}
  \item{I4}{Subsidy for heating devices for households}
  \item{I5}{Record: inventory and information}
  \item{I6}{Dry firewood}
  \item{I7}{Inspection and audit}
  \item{I8}{Improvement and production}
  \item{I9}{Subsidy for firewood}
  \item{I10}{Subsidy for thermal insulation of houses}
  \item{I11}{Education}
  \item{I12}{Subsidy for boilers}
  \item{I13}{R&D, research and development}
  \item{I14}{Subsidy for transportation}
  \item{I15}{Environmental impact assessment system}
  \item{I16}{Mitigation}
}
}
\usage{
CC
}
\description{
We illustrate the use of the foRgotten package with data on public policies
on air pollution from the Valdivia commune, Chile. Studied by Manna, E. M. et al. (2017), where
incentives were considered to promote behavior change in public policy on air pollution.
}
\details{
CC
}
\references{
Manna, E. M., Rojas-Mora, J., & Mondaca-Marino, C. (2017).
Application of the Forgotten Effects Theory for Assessing the Public Policy
on Air Pollution of the Commune of Valdivia, Chile. In
\emph{From Science to Society: New Trends in Environmental Informatics}
(pp. 61-72). Cham: Springer International Publishing.
}
\keyword{datasets}
