% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directEffects.R
\name{directEffects}
\alias{directEffects}
\title{Significant Direct Effects For Complete and Chain Bipartite Graphs}
\usage{
directEffects(
  CC = NULL,
  CE = NULL,
  EE = NULL,
  thr = 0.5,
  conf.level = 0.95,
  reps = 10000,
  delete = FALSE
)
}
\arguments{
\item{CC}{It allows for entering a three-dimensional incidence array, with each submatrix along the z-axis being a square incidence matrix. By default, CC = NULL.}

\item{CE}{It allows for entering a three-dimensional incidence array, with each submatrix along the z-axis being a square incidence matrix (for complete graphs) or a rectangular matrix (for chained bipartite graphs).}

\item{EE}{It allows for entering a three-dimensional incidence array, with each submatrix along the z-axis being a square incidence matrix. By default, EE = NULL.}

\item{thr}{Defines the degree of truth in which incidence is considered significant within the range [0,1]. By default thr = 0.5.}

\item{conf.level}{Defines the confidence level within the range [0,1]. By default, conf.level = 0.95.}

\item{reps}{Defines the number of bootstrap replicates. By default, reps = 10000.}

\item{delete}{Removes the non-significant results from the $DirectEffects set and returns the entered three-dimensional incidence arrays by assigning zeros to the edges whose incidences are significantly lower than thr at the p-value set in the conf.level parameter. By default, delete = FALSE.

If you use the full graph mode, make sure to keep the default values of the CE and EE parameters.}
}
\value{
The function returns a list with subsets of data.
The $DirectEffects subset contains the following values:
 \item{From}{Origin of the incident.}
 \item{To}{Destination of the incident.}
 \item{Mean}{Average incidence.}
 \item{UCI}{Upper Confidence Interval}
 \item{p.value}{The calculated p-value.}

For delete = TRUE with complete graphs, the function returns $Data, the three-dimensional
 matrix entered in the CC parameter, but assigning 0 to the non-significant edges.

For delete = TRUE with chain bipartite graphs, the function returns $CC, $CE, $EE,
the three-dimensional matrices entered in the parameters CC, CE and EE, but assigning
0 to the non-significant edges.
}
\description{
Performs the calculation of the mean incidence, left one-sided confidence interval, and p-value with multiple key informants for complete graphs and chained bipartite graphs. The function allows assigning zeros to the edges whose mean incidences are not significant at the set p-value.
}
\details{
directEffects
}
\examples{
#' # For complete graphs only the CC parameter is used.
# For instance:
directEffects(CE = CE, thr = 0.5, conf.level=0.95, reps = 100, delete = FALSE)
# For chain bipartite graphs the parameters CC, CE and EE are used.
# For instance:
directEffects(CC = CC, CE = CE, EE = EE, thr = 0.5, conf.level=0.95, reps = 100, delete = FALSE)
}
