% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadFiles.R
\name{loadFiles}
\alias{loadFiles}
\title{loadFiles}
\usage{
loadFiles(
  directory,
  sep = ",",
  header = T,
  row.names = 1,
  maxnorm = 10,
  minnorm = 0,
  diagone = T,
  NAvalue = 0.5
)
}
\arguments{
\item{directory}{directory}

\item{sep}{sep}

\item{header}{header}

\item{row.names}{row.names}

\item{maxnorm}{maxnorm}

\item{minnorm}{minnorm}

\item{diagone}{diagone}

\item{NAvalue}{NAvalue}
}
\value{
return
}
\description{
loadFiles
}
